# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: BBYO Connect API
  version: 1.0.0
  summary: Secure closed social network backend for BBYO teens and chapter accounts
  description: |
    Production-oriented backend API for BBYO Connect.

    Security and youth safety features:
    - JWT access tokens + rotating refresh tokens
    - Role-based access control
    - Audit logging
    - Reporting and moderation workflows
    - Language filtering hooks

    Realtime channel (Socket.IO):
    - message.created
    - message.edited
    - message.deleted
    - conversation.updated
    - notification.created
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          description: Optional error details
    ApiEnvelope:
      type: object
      required: [data, meta, error]
      properties:
        data:
          description: Response payload
        meta:
          description: Pagination or extra metadata
        error:
          oneOf:
            - type: "null"
            - $ref: '#/components/schemas/ApiError'
...
    Locale:
      type: string
      enum: [en, es, he, fr]
    RegisterRequest:
      type: object
      required: [accountType, memberId, password]
      properties:
        accountType:
          type: string
          enum: [teen, chapter]
        email:
          type: string
          format: email
        memberId:
          type: string
        password:
          type: string
          minLength: 8
        profile:
          type: object
          additionalProperties: true
    LoginRequest:
      type: object
      required: [memberIdOrEmail, password]
      properties:
        memberIdOrEmail:
          type: string
        password:
          type: string
          minLength: 8
        deviceInfo:
          type: object
          additionalProperties: true
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
        deviceInfo:
          type: object
          additionalProperties: true
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register teen or chapter account
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login with memberId/email and password
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
  /auth/guest:
    post:
      tags: [Auth]
      security: []
      summary: Create guest session
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                locale:
                  $ref: '#/components/schemas/Locale'
      responses:
        '201':
          description: Guest user created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate refresh token and return new access token
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token rotated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...
  /auth/delete-account:
    delete:
      tags: [Auth]
      summary: Soft-delete account and revoke sessions
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...
  /chapter-approvals/request:
    post:
      tags: [Chapter Approvals]
      summary: Create chapter approval request
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /chapter-approvals/{id}/approve:
    patch:
      tags: [Chapter Approvals]
      summary: Approve chapter request
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
      responses:
        '200':
          description: Request approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
  /chapter-approvals/{id}/reject:
    patch:
      tags: [Chapter Approvals]
      summary: Reject chapter request
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
      responses:
        '200':
          description: Request rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope' }
...